Progetto/src/requirements.txt
aiocoap==0.3
alabaster==0.7.12
apparmor==2.12
apt-xapian-index==0.47
apturl==0.5.2
args==0.1.0
asn1crypto==0.24.0
attrs==19.3.0
Babel==2.8.0
backcall==0.1.0
beautifulsoup4==4.6.0
bitarray==1.5.2
bleach==3.1.0
blinker==1.4
Brlapi==0.6.6
certifi==2020.6.20
chardet==3.0.4
clint==0.5.1
command-not-found==0.3
coverage==5.0.3
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
decorator==4.4.1
defer==1.0.6
defusedxml==0.6.0
devscripts===2.17.12ubuntu1.1
distro-info===0.18ubuntu0.18.04.1
docopt==0.6.2
docutils==0.16
entrypoints==0.3
enum34==1.1.10
flake8==3.7.9
fog05==0.2.1
fog05-sdk==0.2.1
gpg==1.10.0
hexdump==3.3
html5lib==0.999999999
httplib2==0.9.2
idna==2.10
imagesize==1.2.0
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
jedi==0.16.0
Jinja2==2.11.2
joblib==0.14.1
json5==0.9.1
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.2
jupyterlab==1.2.6
jupyterlab-server==1.0.6
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
LibAppArmor==2.12
louis==3.5.0
lxml==4.5.2
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.1.1
matplotlib==3.1.3
mccabe==0.6.1
mistune==0.8.4
mock==3.0.5
more-itertools==8.2.0
mutagen==1.44.0
mvar==0.0.1
nbconvert==5.6.1
nbformat==5.0.4
netifaces==0.10.9
nose==1.3.7
notebook==6.0.3
numpy==1.18.1
oauth==1.0.1
oauthlib==2.0.6
olefile==0.45.1
packaging==20.4
paho-mqtt==1.5.0
pandas==1.0.1
pandocfilters==1.4.2
papero==0.2.7
parso==0.6.1
pbr==5.4.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==5.1.0
pluggy==0.13.1
prometheus-client==0.7.1
prompt-toolkit==3.0.3
protobuf==3.0.0
psutil==5.7.2
ptyprocess==0.6.0
py==1.8.1
pyang==2.3.2
pyangbind==0.8.1
pycairo==1.16.2
pycodestyle==2.5.0
pycrypto==2.6.1
pycups==1.9.73
pyflakes==2.1.1
Pygments==2.6.1
pygobject==3.26.1
PyJWT==1.5.3
pylxd==2.2.10
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.4.7
pyRFC3339==1.0
pyrsistent==0.15.7
pytest==5.3.5
python-apt==1.6.5+ubuntu0.3
python-dateutil==2.8.1
python-debian==0.1.32
python-magic==0.4.16
python-xapp==1.0.0
python-xlib==0.20
pytz==2020.1
pyxattr==0.6.0
pyxdg==0.25
PyYAML==3.12
pyzmq==18.1.1
randompy==1.0
regex==2020.7.14
reportlab==3.4.0
requests==2.24.0
requests-oauthlib==0.8.0
requests-toolbelt==0.9.1
requests-unixsocket==0.1.5
scdl==1.6.12
scikit-learn==0.22.1
scipy==1.4.1
SecretStorage==2.3.1
Send2Trash==1.5.0
setproctitle==1.1.10
simplejson==3.13.2
six==1.15.0
snowballstemmer==2.0.0
Sphinx==3.2.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
ssh-import-id==5.7
system-service==0.3
systemd-python==234
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
traitlets==4.3.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
unidiff==0.5.4
urllib3==1.25.10
usb-creator==0.3.3
wadllib==1.3.2
wcwidth==0.1.8
webencodings==0.5.1
ws4py==0.5.1
xkit==0.0.0
yaks==0.3.0.post1
zenoh==0.3.0
zipp==2.2.0
zope.interface==4.3.2
